requirements.txt
boto3
botocore
embedchain[poe,slack]
python-dotenv
slack-bolt
